:sparkles: Add EventBannerSlider component
Presenting the EventBannerSlider component, a React component that showcases event information and includes a banner slider for displaying movies. It fetches movie data and presents the information along with optional date and description. The component features a loading state with a simulated delay and handles the rendering of content and skeleton loader accordingly.